+++ src/gov/epd/index.js
@@ -1,0 +1,5 @@
+module.exports = {
+    searchApi: require("./api"),
+    searchAqhi: require("./aqhi"),
+    latest: require("./latest"),
+}