+++ src/gov/epd/latest/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+    searchAqhi: require("./aqhi"),
+}